:bug: make Mask<T, U> only accept methods on T
Mask<T, U> used to accept any type U - if U didn't intersect with T
at all the type checker would let *everything* through, essentially
failing silently. This is not a bug in Typescript; it's how union
types are supposed to work. The right thing to do here is not to use
a union type at all to perform the masking, that way providing methods
in U that aren't on T will simply make it so that we can't assign T
to Mask<T, U>, which will definitely trigger failures.